+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Only copy the static files that changed, not the whole directory in `gutenberg serve`
 - Use summary if available in RSS feed
 - Add tables and footnotes support in markdown
+- Add previous/previous_in_section/next/next_in_section/summary to `Page` 
 
 ## 0.0.3 (2017-04-05)
 - Add some colours in console